+++ bdi_game/src/algorithms/search/algorithms/mod.rs
@@ -1,2 +1,4 @@
 pub mod heuristic_search;
 pub mod naive_search;
+mod ongoing_search;
+mod a_star_search;
